requirements.txt
pytest
pytest-rerunfailures
requests
PyYAML
allure-pytest
playwright